Optimize querying by limiting the amount of results
By default the query is now limited to 10 items.
However if the bot notices that there might be more than 10 new items,
it gets all of them.

This saves a lot of time in all the normal use cases (when there are
less than 10 new items per update per vahti)